Add files via upload
This Excel folder includes two spreadsheets, one named 'SYR' and one named 'SPI-SPEI'.
The 'SYR' spreadsheet contains wheat crop yield data (ton/(ha y)) for the four provinces of the Abruzzo region, in central Italy, provided by the Italian National Institute of Statistics (ISTAT), recorded from the year 1952 to 2014 The parameters of the trendlines are also provided, calculated on the 1952-1982 and 1953-2014 time intervals, the deviations (or residuals) between the recorded yield and that predicted by the trendline, and the standardized residuals (SYR).
The SPI-SPEI sheet contains the values ​​of the SPI and SPEI indices at one, three and six months, estimated in different months of the year, for each of the four provinces in Abruzzo. An exception is the province of L'Aquila, for which the SPEI indices were not estimated but only the SPI indices. SPI and SPEI indices have been estimated based on temperature and precipitation time series, recorded in 37 stations distributed throughout the regional territory, in the time interval from 1952 to 2014.